optimize dtos by including create and return dtos

+++ wte-backend/src/main/java/net/summer23project/wtebackend/dto/IngredientNutrientAmountDto.java
@@ -15,5 +15,5 @@ import lombok.Setter;
 public class IngredientNutrientAmountDto {
     private String ingredientName;
     private String nutrientName;
-    private int nutrientAmount;
+    private double nutrientAmount;
 }
